+++ autosubmit/statistics/jobs_stat.py
@@ -24,11 +24,30 @@ from log.log import Log
 
 
 def _estimate_requested_nodes(nodes, processors, tasks, processors_per_node) -> int:
+    """Estimates the number of requested nodes.
+
+    In the past we had ``ZeroDivisionError`` due to the number of ``tasks`` being set
+    to ``0`` by default in the ``Job`` class. If that changes, we can remove the checks
+    here.
+
+    If ``nodes`` is provided and valid, that's returned immediately.
+
+    If ``tasks`` is provided and valid, and ``tasks`` is greater than ``0`` (to prevent
+    the ``ZeroDivisionError``) then we return the ceiling value of ``processors``
+    divided by ``tasks``.
+
+    If ``processors_per_node`` is provided and valid, and ``processors_per_node`` is
+    greater than zero (``ZeroDivisionError``), and ``processors_per_node`` is not
+    greater than ``processors``, then we return the ceiling value of ``processors``
+    divided by the number of ``processors_per_node``.
+
+    Else, we return ``1``.
+    """
     if str(nodes).isdigit():
         return int(nodes)
-    elif str(tasks).isdigit():
+    elif str(tasks).isdigit() and int(tasks) > 0:
         return ceil(int(processors) / int(tasks))
-    elif str(processors_per_node).isdigit() and int(processors) > int(processors_per_node):
+    elif str(processors_per_node).isdigit() and 0 < int(processors_per_node) < int(processors):
         return ceil(int(processors) / int(processors_per_node))
     else:
         return 1